+++ src/model/Clothing.java
@@ -1,11 +1,13 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Clothing extends Product {
 
 	private String size;
 	private String colour;
 	
-	public Clothing(String id, String name, double puchasePrice, double salesPrice, String countryOfOrigin,
+	public Clothing(String id, String name, BigDecimal puchasePrice, BigDecimal salesPrice, String countryOfOrigin,
 			int minStock, String colour, String size) {
 		super(id, name, puchasePrice, salesPrice, countryOfOrigin, minStock);
 		this.size = size;
