+++ src/model/Equipment.java
@@ -1,11 +1,13 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Equipment extends Product {
 
 	private String type;
 	private String description;
 	
-	public Equipment(String id, String name, double puchasePrice, double salesPrice, String countryOfOrigin,
+	public Equipment(String id, String name, BigDecimal puchasePrice, BigDecimal salesPrice, String countryOfOrigin,
 			int minStock, String type, String description) {
 		super(id, name, puchasePrice, salesPrice, countryOfOrigin, minStock);
 		this.type = type;
